refactor(bookings): réécrit convert_parking_to_van — Parking est une Room, pas un Space
Correction après vérification sur copie prod : « Parking » (PKG) est la chambre
(Room code PKG, name Parking) du gîte Espace camping-cars, pas un Space. Les
réservations à convertir sont des Booking d hébergement dont les Reservation
pointent uniquement cette chambre (16 bookings PKG-purs mesurés en prod).

- Périmètre : Booking vivants dont TOUTES les Reservation vivantes pointent la
  Room PKG (résolue par code, repli name Parking — jamais par id en dur) ;
  mixtes/empty/no_stay rapportés et non touchés
- Conversion : VanBooking 1 véh., from/to copiés du Booking (sémantique nuits
  déjà portée), price_cents conservé ; garde-fou date_divergence vs min/max+1
  des Reservation (informatif)
- soft-delete du StayItem du Booking puis du Booking (cascade soft-delete des
  Reservation → chambre libérée), assertion total du séjour inchangé
- volet SpaceBooking retiré (lib/tasks/spaces.rake supprimé)
- transaction requires_new, DRY-RUN par défaut / APPLY=1, idempotent
- specs réécrites (4 ex.) ; suite complète verte (843 ex., 0 échec)
